requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
aws-shell==0.2.2
awscli==1.32.69
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.69
botocore==1.34.69
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
configobj==5.0.8
crashtest==0.4.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==5.0.3
docutils==0.16
dulwich==0.21.7
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
image==1.5.33
importlib_metadata==7.1.0
install==1.3.5
installer==0.7.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==24.3.1
langchain==0.1.13
langchain-cli==0.0.21
langchain-community==0.0.29
langchain-core==0.1.33
langchain-text-splitters==0.0.1
langserve==0.0.51
langsmith==0.1.31
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==1.0.18
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
pyproject_hooks==1.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.7.2
s3transfer==0.10.1
selenium==4.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.36.3
tenacity==8.2.3
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
trove-classifiers==2024.3.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.23.2
virtualenv==20.25.1
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
xattr==1.1.0
yarl==1.9.4
zipp==3.18.1
